Use u128 for account balances
