+++ bcfs/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "bcfs"
-version = "0.2.2"
+version = "0.2.3"
 license = "Apache-2.0"
 authors = ["Oasis Labs <feedback@oasislabs.com>"]
 edition = "2018"
