+++ blockchain-traits/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "blockchain-traits"
-version = "0.2.2"
+version = "0.2.3"
 authors = ["Oasis Labs <feedback@oasislabs.com>"]
 edition = "2018"
 license = "Apache-2.0"
