+++ memchain/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "memchain"
-version = "0.2.2"
+version = "0.2.3"
 authors = ["Oasis Labs <feedback@oasislabs.com>"]
 edition = "2018"
 license = "Apache-2.0"
