+++ oasis-test/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "oasis-test"
-version = "0.2.0"
+version = "0.2.1"
 license = "Apache-2.0"
 description = "A testing framework for Oasis executables."
 edition = "2018"
