+++ oasis-types/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "oasis-types"
-version = "0.2.0"
+version = "0.2.1"
 authors = ["Oasis Labs <feedback@oasislabs.com>"]
 edition = "2018"
 license = "Apache-2.0"
