+++ oasis-types/src/lib.rs
@@ -1,117 +1,9 @@
 #[macro_use]
 extern crate serde;
 
-/// A 160-bit little-endian hash address type.
-#[derive(Clone, Copy, Default, PartialEq, Eq, PartialOrd, Ord, Debug, Hash, Serialize)]
-#[repr(C)]
-pub struct Address(pub [u8; 20]);
+mod address;
 
-impl Address {
-    /// Creates an `Address` from a little-endian byte array.
-    pub unsafe fn from_raw(bytes: *const u8) -> Self {
-        let mut addr = Self::default();
-        addr.0
-            .copy_from_slice(std::slice::from_raw_parts(bytes, 20));
-        addr
-    }
-
-    pub fn as_ptr(&self) -> *const u8 {
-        self.0.as_ptr()
-    }
-
-    pub const fn len() -> usize {
-        std::mem::size_of::<Self>()
-    }
-
-    pub fn path_repr(&self) -> String {
-        hex::encode(self)
-    }
-}
-
-impl AsRef<[u8]> for Address {
-    fn as_ref(&self) -> &[u8] {
-        &self.0
-    }
-}
-
-impl std::str::FromStr for Address {
-    type Err = hex::FromHexError;
-    fn from_str(s: &str) -> Result<Self, Self::Err> {
-        let bytes: Vec<u8> = hex::decode(s)?;
-        if bytes.len() != Address::len() {
-            return Err(hex::FromHexError::InvalidStringLength);
-        }
-        let mut addr = Self::default();
-        addr.0.copy_from_slice(&bytes);
-        Ok(addr)
-    }
-}
-
-impl std::fmt::Display for Address {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        write!(f, "0x{}", hex::encode(self.0))
-    }
-}
-
-impl<'de> serde::de::Deserialize<'de> for Address {
-    fn deserialize<D>(deserializer: D) -> Result<Self, D::Error>
-    where
-        D: serde::de::Deserializer<'de>,
-    {
-        use serde::de;
-
-        const EXPECTATION: &str = "20 bytes";
-
-        struct AddressVisitor;
-        impl<'de> de::Visitor<'de> for AddressVisitor {
-            type Value = Address;
-
-            fn expecting(&self, f: &mut std::fmt::Formatter) -> std::fmt::Result {
-                f.write_str(EXPECTATION)
-            }
-
-            fn visit_seq<V>(self, mut seq: V) -> Result<Self::Value, V::Error>
-            where
-                V: de::SeqAccess<'de>,
-            {
-                let mut arr = [0; Self::Value::len()];
-
-                if let Some(len) = seq.size_hint() {
-                    if len != arr.len() {
-                        return Err(de::Error::invalid_length(len, &EXPECTATION));
-                    }
-                }
-
-                let mut i = 0;
-                loop {
-                    match seq.next_element()? {
-                        Some(el) if i < arr.len() => arr[i] = el,
-                        None if i == arr.len() => break,
-                        _ => return Err(de::Error::invalid_length(i, &EXPECTATION)),
-                    }
-                    i += 1;
-                }
-
-                Ok(Address(arr))
-            }
-
-            fn visit_bytes<E>(self, value: &[u8]) -> Result<Self::Value, E>
-            where
-                E: de::Error,
-            {
-                let mut arr = [0; std::mem::size_of::<Self::Value>()];
-                if value.len() == arr.len() {
-                    arr.copy_from_slice(value);
-                    Ok(Address(arr))
-                } else {
-                    Err(de::Error::invalid_length(value.len(), &EXPECTATION))
-                }
-            }
-        }
-
-        deserializer.deserialize_any(AddressVisitor)
-    }
-}
+pub use address::Address;
 
 #[repr(C)]
 #[derive(PartialEq, Eq)]
@@ -128,7 +20,7 @@ impl ExtStatusCode {
 
 #[derive(Clone, Default, Debug)]
 pub struct AccountMeta {
-    pub balance: u64,
+    pub balance: u128,
     pub expiry: Option<std::time::Duration>,
 }
